two undefined shifts the sanitizer leg found on its first run, and t_fetch's manifest
The ASan/UBSan matrix job ran for the first time on the v1.43 recut and halted
seven tests on two sites that every other build had been executing for years:

  src/format4.c:190  left shift of 1 by 31 places cannot be represented in 'int'
  src/format2.c:146  shift exponent 32 is too large for 32-bit type 'int'

The first is `1<<31` where the line above it already wrote `1ul<<31`; now 1u.
The second is worse than it looks: f2_get_uint64 assembled its return value
from bytes with `d[j] << (8*j)` and d[j] promotes to int, so at j == 3 it
shifted into the sign bit and at j == 4 by the full width -- both undefined --
and the function only ever returned the uint64_t it promises by the compiler
happening to do the obvious thing. It widens before shifting now. Behaviour is
unchanged where it was defined: state tracks round-trip and unpack output is
byte-identical to v1.41.

Also: the recipe's test phase copies test/ but t_fetch builds its loopback
mirror from a cached manifest under tools/registry/sums, so it found none and
failed. That directory rides along now.

+++ src/format2.c
@@ -143,7 +143,10 @@ uint64_t f2_get_uint64(cdata_t *c, uint64_t i) {
   f2_aux_t *aux = (f2_aux_t*) c->aux;
   uint8_t *d = aux->data + c->unit*i;
   uint64_t value = 0;
-  for (uint8_t j=0; j<c->unit; ++j) value |= (d[j] << (8*j));
+  /* Widen BEFORE shifting. d[j] promotes to int, so at j == 4 this shifted a
+   * 32-bit value by 32 (undefined) and at j == 3 into the sign bit (also
+   * undefined); the function promises a uint64_t and only got one by luck. */
+  for (uint8_t j=0; j<c->unit; ++j) value |= ((uint64_t) d[j] << (8*j));
   return value;
 }
